cleanup(storage)!: reorganize errors and builders
Move all the errors to a module. Refactor the `KeyAes256` helper to the
builders, but maybe that is not the best place either.
